requirements.txt
streamlit==1.32.0
psycopg[binary]==3.1.14
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
